% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeevents.R
\name{filter_le}
\alias{filter_le}
\title{Filtering Life Events by Age and Influence}
\usage{
filter_le(data, age.filter, max.age, include.last.year = T, influence.filter,
  min.influence, reliable)
}
\arguments{
\item{data}{A data frame containing the life events data}

\item{age.filter}{A logical. If age.filter = T, life events will be filtered
according to the max.age and include.last.year settings. Missing age information
will set life event to missing.}

\item{max.age}{A string. Name of the column with the maximum allowable age. Since
the age at which a life event happened is only available at whole years,
this by default can also include events, which happened after the allowable
age, if include.last.year = T.}

\item{influence.filter}{A logical. If influence.filter=T, life events will
be filtered according to the min.influence setting. Missing influence
information will set life events to missing.}

\item{min.influence}{A numeric value. The minimum amount of influence
at the time of interview.
\enumerate{
 \item 'no influence / geen'
 \item 'a little influence / een beetje'
 \item 'moderate influence / tamelijk veel'
 \item 'a lot of influence / heel veel'
}}

\item{reliable}{A logical. Indicates whether life events should be omitted,
when they were judged to be unreliable, e.g. when language difficulties
were noted during the interview. reliable = T omits unreliable life events,
reliable = F keeps them. Missing reliability information is assumed to
be reliable.}
}
\value{
A data.frame containing information whether life events occured.
}
\description{
\code{filter_le} filters life events by a specified age and influence at
time of interview. By default life events occuring at the same age year as
the specified age filter are kept. Life events passing the filteres are
indicated by 1 and are otherwise 0.
}

