% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_help.R
\docType{package}
\name{LifeEvents}
\alias{LifeEvents}
\alias{LifeEvents-package}
\title{LifeEvents}
\description{
This package helps working with the Generation R life events data.
It contains functions to filter life events depending on age of
occurence and influence, and to compute the number of life events
experienced.
}
\details{
\code{\link{filter_le}} filters life events by a specified age and influence at
time of interview.

\code{\link{sum_le}} computes the number of life events experienced.
}

