% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeevents.R
\name{sum_le}
\alias{sum_le}
\title{Compute number of life events}
\usage{
sum_le(data, include.open, na.rm)
}
\arguments{
\item{data}{A data.frame, which includes information, whether life events
occured or not.}

\item{include.open}{A logical. include.open = T will include open life
events in calculations.}

\item{na.rm}{A logical. na.rm=T will treat missing life events as 0.
na.rm = F will set number of life events to missing in presence of missing
life events.}
}
\description{
\code{sum_le} computes the number of life events experienced. Options
include to compute the number omitting open life events, as well as
how to handle missing values.
}

