% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeevents.R
\name{sum_le}
\alias{sum_le}
\title{Compute number of life events}
\usage{
sum_le(data, range = c(1:24), na.rm = F)
}
\arguments{
\item{data}{A matrix, which includes information, whether life evetns
occured (1) or not (0).}

\item{range}{A numeric vector. Indicates which life events should be taken
into account. The default is c(1:24), which stands for all
closed events.}

\item{na.rm}{A logical. na.rm=T will treat missing life events as 0.
na.rm = F will set number of life events to missing in presence of missing
life events.}
}
\description{
\code{sum_le} computes the number of life events experienced. Options
include to only take into account a subset of LE, as well as
how to handle missing values. Completely missing rows are set to missing.
}

